resoling of global variables are now moved to GlobalVariableResolver, a Utility class, that prepared for a future update, when variables are reworked completely
