+++ src/main/java/org/betonquest/betonquest/api/config/quest/QuestPackage.java
@@ -68,6 +68,14 @@ public interface QuestPackage {
     @Deprecated
     String getRawString(String address);
 
+    /**
+     * Resolves all global variables in the given string.
+     *
+     * @param input the string to resolve
+     * @return the resolved string
+     * @deprecated instead use
+     * {@link org.betonquest.betonquest.variables.GlobalVariableResolver#resolveGlobalVariables(QuestPackage, String)}.
+     */
     @Deprecated
     String subst(String input);
