+++ src/main/java/org/betonquest/betonquest/modules/config/quest/QuestPackageImpl.java
@@ -4,18 +4,14 @@ import lombok.CustomLog;
 import org.betonquest.betonquest.api.bukkit.config.custom.multi.MultiConfiguration;
 import org.betonquest.betonquest.api.config.ConfigAccessor;
 import org.betonquest.betonquest.api.config.quest.QuestPackage;
-import org.betonquest.betonquest.exceptions.ObjectNotFoundException;
-import org.betonquest.betonquest.id.GlobalVariableID;
 import org.betonquest.betonquest.utils.Utils;
+import org.betonquest.betonquest.variables.GlobalVariableResolver;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.InvalidConfigurationException;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.List;
-import java.util.Locale;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * This {@link QuestPackageImpl} represents all functionality based on a {@link Quest}.
@@ -73,127 +69,8 @@ public class QuestPackageImpl extends QuestTemplate implements QuestPackage {
     }
 
     @Override
-    @SuppressWarnings({"PMD.CognitiveComplexity", "PMD.CyclomaticComplexity", "PMD.NPathComplexity", "PMD.NcssCount"})
     public String subst(final String input) {
-        if (input == null) {
-            return null;
-        }
-        String variableInput = input;
-
-        final Pattern globalVariableRegex = Pattern.compile("\\$([^ $\\s]+)\\$");
-        while (true) {
-            final Matcher matcher = globalVariableRegex.matcher(variableInput);
-            if (!matcher.find()) {
-                break;
-            }
-            final String varName = matcher.group(1);
-            final String varVal;
-            if ("this".equals(varName)) {
-                varVal = getQuestPath();
-            } else {
-                final GlobalVariableID variableID;
-                try {
-                    variableID = new GlobalVariableID(this, varName);
-                } catch (final ObjectNotFoundException e) {
-                    LOG.warn(this, e.getMessage(), e);
-                    return variableInput;
-                }
-                final String varRaw = variableID.getPackage().getConfig().getString("variables." + variableID.getBaseID());
-                if (varRaw == null) {
-                    LOG.warn(this, String.format("Variable %s not defined in package %s", variableID.getBaseID(), variableID.getPackage().getQuestPath()));
-                    return variableInput;
-                }
-                varVal = resolve(variableID.getPackage(), varRaw);
-            }
-
-            if (varVal
-                    .matches("^\\$[a-zA-Z\\d]+\\$->\\(-?\\d+\\.?\\d*;-?\\d+\\.?\\d*;-?\\d+\\.?\\d*\\)$")) {
-                final String innerVarName = varVal.substring(1, varVal.indexOf('$', 2));
-                final String innerVarVal = getConfig().getString("variables." + innerVarName);
-                if (innerVarVal == null) {
-                    LOG.warn(this, String.format("Location variable %s is not defined, in variable %s, package %s.",
-                            innerVarName, varName, getQuestPath()));
-                    return variableInput;
-                }
-
-                if (!innerVarVal.matches("^-?\\d+;-?\\d+;-?\\d+;.+$")) {
-                    LOG.warn(this,
-                            String.format("Inner variable %s is not valid location, in variable %s, package %s.",
-                                    innerVarName, varName, getQuestPath()));
-                    return variableInput;
-                }
-
-                final double locX;
-                final double locY;
-                final double locZ;
-                final String rest;
-                try {
-                    final int offset1 = innerVarVal.indexOf(';');
-                    locX = Double.parseDouble(innerVarVal.substring(0, offset1));
-                    final int offset2 = innerVarVal.indexOf(';', offset1 + 1);
-                    locY = Double.parseDouble(innerVarVal.substring(offset1 + 1, offset2));
-                    final int offset3 = innerVarVal.indexOf(';', offset2 + 1);
-                    locZ = Double.parseDouble(innerVarVal.substring(offset2 + 1, offset3));
-                    rest = innerVarVal.substring(offset3);
-                } catch (final NumberFormatException e) {
-                    LOG.warn(this, String.format(
-                            "Could not parse coordinates in inner variable %s in variable %s in package %s",
-                            innerVarName, varName, getQuestPath()), e);
-                    return variableInput;
-                }
-                final double vecLocX;
-                final double vecLocY;
-                final double vecLocZ;
-                try {
-                    final int offset1 = varVal.indexOf('(');
-                    final int offset2 = varVal.indexOf(';');
-                    final int offset3 = varVal.indexOf(';', offset2 + 1);
-                    final int offset4 = varVal.indexOf(')');
-                    vecLocX = Double.parseDouble(varVal.substring(offset1 + 1, offset2));
-                    vecLocY = Double.parseDouble(varVal.substring(offset2 + 1, offset3));
-                    vecLocZ = Double.parseDouble(varVal.substring(offset3 + 1, offset4));
-                } catch (final NumberFormatException e) {
-                    LOG.warn(this, String.format("Could not parse vector in location variable %s in package %s",
-                            varName, getQuestPath()), e);
-                    return variableInput;
-                }
-                final double locationX = locX + vecLocX;
-                final double locationY = locY + vecLocY;
-                final double locationZ = locZ + vecLocZ;
-                variableInput = variableInput.replace("$" + varName + "$", String.format(Locale.US, "%.2f;%.2f;%.2f%s", locationX, locationY, locationZ, rest));
-            } else {
-                variableInput = variableInput.replace("$" + varName + "$", varVal);
-            }
-        }
-
-        return variableInput;
-    }
-
-    private static String resolve(final QuestPackage pack, final String input) {
-        if (input == null) {
-            return null;
-        }
-        final Pattern globalVariableRegex = Pattern.compile("\\$([^ $\\s]+)\\$");
-        final Matcher matcher = globalVariableRegex.matcher(input);
-        final StringBuilder builder = new StringBuilder();
-        while (matcher.find()) {
-            final String varName = matcher.group(1);
-            final String varVal;
-            if ("this".equals(varName)) {
-                varVal = pack.getQuestPath();
-            } else {
-                try {
-                    final GlobalVariableID variableID = new GlobalVariableID(pack, varName);
-                    varVal = "$" + variableID.getPackage().getQuestPath() + "." + variableID.getBaseID() + "$";
-                } catch (final ObjectNotFoundException e) {
-                    LOG.warn(pack, e.getMessage(), e);
-                    return input;
-                }
-            }
-            matcher.appendReplacement(builder, Matcher.quoteReplacement(varVal));
-        }
-        matcher.appendTail(builder);
-        return builder.toString();
+        return GlobalVariableResolver.resolveGlobalVariables(this, input);
     }
 
     @Override
@@ -211,7 +88,7 @@ public class QuestPackageImpl extends QuestTemplate implements QuestPackage {
             return value;
         }
 
-        return subst(value);
+        return GlobalVariableResolver.resolveGlobalVariables(this, value);
     }
 
     @Override
